fix: resolve 14 code-review issues across Core/Differential/Bowl/Drivelution/Extension
## Bug fixes

- **Brotli patch detection**: add BrotliFormatVersion (0x02) to format detection
  and decompression switch in BsdiffDiffer (#512)
- **ReadFileWithBudget**: trim buffer to actual bytes read to prevent zero-padded
  data corrupting patch computation (#512)
- **BsdiffDiffer.Search**: guard against sentinel -1 values from Split() that
  would cause IndexOutOfRangeException (#512)
- **Bowl EnsureDirectory**: stop deleting the fail directory on every Prepare()
  call — preserves crash diagnostics from prior sessions (#512)
- **FileNode.Equals**: return false for incompatible types instead of throwing
  ArgumentException, conforming to the IEquatable contract (#512)
- **EventManager.Dispatch**: snapshot the delegate invocation list before
  enumerating to avoid race with concurrent AddListener/RemoveListener (#512)
- **DiffPipeline**: add 'using' to both SemaphoreSlim instances to fix kernel
  handle leaks (#512)
- **FileTree.Compare**: guard all 'node0' access paths with null-conditional
  operators to prevent NullReferenceException (#512)
- **ClientStrategy.CheckFail**: use Version.TryParse instead of bare
  new Version() to avoid crashing on malformed version strings (#512)
- **GeneralTracer**: wrap InitializeFileListener with lock to prevent duplicate
  TextTraceListener creation on concurrent day-rollover (#512)
- **ProcessRunner**: cancel the Task.Delay timer via CancellationTokenSource
  when the process exits first, preventing timer leak (#512)
- **OssStrategy**: resolve upgrade app name and search pattern based on
  RuntimeInformation.IsOSPlatform instead of hardcoding .exe (#512)
- **VersionComparer**: use long instead of int for major/minor/patch to
  prevent overflow on extremely large version numbers (#512)
- **ExtensionGeneralHost**: move name extraction before lifecycle hook calls
  and use extensionName (not file path) as Id in ExtensionMetadata (#512)

## Test coverage

- BsdiffDifferTests: 6 new edge-case tests (single-byte, two-byte, repeating
  patterns, single-byte-change round-trips)
- VersionComparerTests: 3 new large-number overflow tests
- FileTreeTests: 6 new tests for null node0, imbalanced trees, matching trees
- FileNodeTests: update Equals_NonFileNodeType test to match new behaviour

All 1864 tests pass across 5 test suites (0 failures).

+++ src/c#/GeneralUpdate.Drivelution/Core/Utilities/VersionComparer.cs
@@ -105,9 +105,9 @@ public static class VersionComparer
 
         return new SemVerInfo
         {
-            Major = int.Parse(match.Groups["major"].Value),
-            Minor = int.Parse(match.Groups["minor"].Value),
-            Patch = int.Parse(match.Groups["patch"].Value),
+            Major = long.Parse(match.Groups["major"].Value),
+            Minor = long.Parse(match.Groups["minor"].Value),
+            Patch = long.Parse(match.Groups["patch"].Value),
             Prerelease = match.Groups["prerelease"].Value,
             BuildMetadata = match.Groups["buildmetadata"].Value
         };
@@ -122,8 +122,8 @@ public static class VersionComparer
 
         for (int i = 0; i < minLength; i++)
         {
-            var isNum1 = int.TryParse(parts1[i], out int num1);
-            var isNum2 = int.TryParse(parts2[i], out int num2);
+            var isNum1 = long.TryParse(parts1[i], out long num1);
+            var isNum2 = long.TryParse(parts2[i], out long num2);
 
             if (isNum1 && isNum2)
             {
@@ -152,9 +152,9 @@ public static class VersionComparer
 
     private class SemVerInfo
     {
-        public int Major { get; set; }
-        public int Minor { get; set; }
-        public int Patch { get; set; }
+        public long Major { get; set; }
+        public long Minor { get; set; }
+        public long Patch { get; set; }
         public string Prerelease { get; set; } = string.Empty;
         public string BuildMetadata { get; set; } = string.Empty;
     }
